preview api

+++ src/main/java/com/example/demo/controller/MediaController.java
@@ -1,18 +1,22 @@
 package com.example.demo.controller;
 
 import com.example.demo.exception.FileNotFoundException;
-import com.example.demo.model.FileMetadata;
 import com.example.demo.service.MediaService;
-import com.example.demo.util.FileConverter;
+import com.example.demo.util.DemoUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.awt.*;
 import java.io.File;
+import java.net.URI;
+import java.util.HashMap;
+import java.util.Map;
 
 @RestController
 public class MediaController {
@@ -31,7 +35,7 @@ public class MediaController {
         }
         try {
             logger.info("Image upload started, fileName: {}, user: {}", inputFile.getName(), userName);
-            File convertedFile = FileConverter.convertToFile(inputFile);
+            File convertedFile = DemoUtil.convertToFile(inputFile);
             mediaService.uploadFile(convertedFile, userName);
         } catch(Exception e) {
             logger.error(e.getMessage(), e);
@@ -46,6 +50,9 @@ public class MediaController {
                                          @PathVariable("fileId") String fileId) {
 
         logger.info("Image delete process started, user {}, file {}", userName, fileId);
+        if(StringUtils.isEmpty(fileId)) {
+            return new ResponseEntity<>("File Id is missing", HttpStatus.BAD_REQUEST);
+        }
         try {
             mediaService.deleteFile(userName, fileId);
         } catch (Exception e) {
@@ -53,5 +60,26 @@ public class MediaController {
             return new ResponseEntity(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
         return new ResponseEntity<>("File deleted successfully", HttpStatus.OK);
+    }
+
+    @GetMapping("/image/delete/{fileId}")
+    public ResponseEntity<Object> viewImage(@RequestParam("userName") String userName,
+                                         @PathVariable("fileId") String fileId) {
+
+        logger.info("Image view process started, user {}, file {}", userName, fileId);
+        if(StringUtils.isEmpty(fileId)) {
+            return new ResponseEntity<>("File Id is missing", HttpStatus.BAD_REQUEST);
+        }
+        Map<String, String> response = null;
+        try {
+            String url = mediaService.viewFile(userName, fileId);
+            response = new HashMap<>();
+            response.put("previewLink", url);
+        } catch (FileNotFoundException e) {
+            logger.error(e.getMessage(), e);
+            return new ResponseEntity(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
